// src-ui/core/ipc/generated/debug/DebugReceiverState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * 事件接收器的连接状态
 */
export type DebugReceiverState = { "state": "connecting" } | { "state": "connected" } | { "state": "reconnecting", attempt: number, retry_in_ms: number, } | { "state": "stopped", reason: string, };